SDGC-99
Crear vista lista de reclamaciones

+++ src/main/java/datos/entity/reclamaciones/Reclamacion.java
@@ -31,7 +31,7 @@ public class Reclamacion {
     private int codigoEmpresaDestino;
 
     @Column(name = "codigo_de_paso")
-    private int codigoDePaso;
+    private long codigoDePaso;
 
     @Column(name = "codigo_del_proceso")
     private String codigoDelProceso;
